+++ logtape/config.ts
@@ -71,14 +71,6 @@ export interface LoggerConfig<
   filters?: TFilterId[];
 
   /**
-   * The log level to filter by.  If `null`, the logger will reject all
-   * records.
-   * @deprecated Use `filters` instead for backward compatibility, or use
-   *             `lowestLevel` for less-misleading behavior.
-   */
-  level?: LogLevel | null;
-
-  /**
    * The lowest log level to accept.  If `null`, the logger will reject all
    * records.
    * @since 0.8.0
@@ -241,7 +233,6 @@ function configureInternal<
   currentConfig = config;
 
   let metaConfigured = false;
-  let levelUsed = false;
   const configuredCategories = new Set<string>();
 
   for (const cfg of config.loggers) {
@@ -272,10 +263,6 @@ function configureInternal<
     logger.parentSinks = cfg.parentSinks ?? "inherit";
     if (cfg.lowestLevel !== undefined) {
       logger.lowestLevel = cfg.lowestLevel;
-    }
-    if (cfg.level !== undefined) {
-      levelUsed = true;
-      logger.filters.push(toFilter(cfg.level));
     }
     for (const filterId of cfg.filters ?? []) {
       const filter = config.filters?.[filterId];
@@ -339,14 +326,6 @@ function configureInternal<
       "<https://logtape.org/manual/categories#meta-logger>.",
     { metaLoggerCategory: ["logtape", "meta"], dismissLevel: "info" },
   );
-
-  if (levelUsed) {
-    meta.warn(
-      "The level option is deprecated in favor of lowestLevel option.  " +
-        "Please update your configuration.  See also " +
-        "<https://logtape.org/manual/levels#configuring-severity-levels>.",
-    );
-  }
 }
 
 /**